outline compare code
