+++ packages/web/src/index.ts
@@ -1,18 +1,26 @@
-import * as element from "./element";
+import * as compare from "./compare";
 import * as event from "./event";
-import * as find from "./find";
 import * as lang from "./lang";
-import * as match from "./match";
-import { Match } from "./match";
 import { Recorder } from "./Recorder";
 import * as select from "./select";
 import { scroll } from "./scroll";
 import * as wait from "./wait";
 import * as xpath from "./xpath";
 
+// TODO remove
+import * as elementToKeep from "./element";
+import * as elementToRemove from "./element.remove";
+const element = { ...elementToKeep, ...elementToRemove };
+import * as findToRemove from "./find.remove";
+import * as findToKeep from "./find";
+const find = { ...findToKeep, ...findToRemove };
+import * as match from "./match.remove";
+import { Match } from "./match.remove";
+
 export type Match = Match;
 
 // export the isomorphic (node & browser) module for node
+const { compareAttributes, compareContent, compareDoc, parseHtml } = compare;
 const { isKeyEvent, isPasteEvent, isTypeEvent } = event;
 const {
   compareArrays,
@@ -24,21 +32,28 @@ const {
 const { isNil } = lang;
 const { sleep, waitFor, waitUntil } = wait;
 export {
-  compareArrays,
-  compareDescriptorKey,
-  compareDescriptors,
-  countPresentKeys,
+  compareAttributes,
+  compareContent,
+  compareDoc,
   isKeyEvent,
   isNil,
   isPasteEvent,
   isTypeEvent,
+  parseHtml,
   sleep,
   waitFor,
-  waitUntil
+  waitUntil,
+  // remove
+  compareArrays,
+  compareDescriptorKey,
+  compareDescriptors,
+  countPresentKeys
+  // remove
 };
 
 // export the web module for the browser
 const webExports = {
+  compare,
   element,
   event,
   find,
